pyproject.toml
[tool.poetry]
name = "desafio-192-cf"
version = "0.1.0"
description = ""
authors = ["Ícaro Abreu <oicaroabreu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-aws-ec2 = "^1.204.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
